% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_degscore.R
\name{degscore}
\alias{degscore}
\title{Get DEG Score.}
\usage{
degscore(
  subtypes = NULL,
  expression_data = swog_expressions,
  this_metadata = swog_meta,
  exclude = NULL,
  plot = FALSE,
  out_path = NULL,
  write_degs = FALSE,
  this_variable = "LundTax2023_simple",
  sample_id_col = "ALTPATID",
  return_this = "degs"
)
}
\arguments{
\item{subtypes}{Optional parameter.Allows the user to subset samples to a specific subtype of interest. To be used together with `this_variable`.}

\item{expression_data}{Required. A data frame with expression values of interest.}

\item{this_metadata}{Required. A data frame with metadata for the samples in the `expression_data` object.}

\item{exclude}{Optional, Lets the user exclude samples of interest. Should be a vector with sample IDs available in the metadata.}

\item{plot}{Boolean parameter, set to TRUE to output plot. Still WIP, not yet available...}

\item{out_path}{Required if `write_degs = TRUE`, target path to exported DEGs, otherwise disregarded.}

\item{write_degs}{Boolean parameter, default is set to FALSE. Set to TRUE for exporting DEGs, if set to TRUE, `out_path` also needs to be specified.}

\item{this_variable}{Required if `subtypes` is defined. Should be the name of the column in the metadata holding subtype classification. Otherwise not needed.}

\item{sample_id_col}{Required parameter, should be the name of the column in the metadata data frame with sample IDs.}

\item{return_this}{Decides what output is returned, could be one of the following; sample_ids, metadata, expressions, eset, design_matrix, degs, all, and nothing}
}
\value{
Data frame with DEGs as well as ranked score.
}
\description{
Return a DEG object with score for each gene.
}
\details{
See parameter descriptions for more info, still wip.
}
\examples{
#load packages
library(limma)
library(Biobase)

## 1. Read and format data
#read expression data
load("C:/Users/matts/Desktop/projects/spp1/SWOG_GEX_ProteinCoding_without_duplica.Rdata", )

#convert to data frame
swog_expressions = as.data.frame(SWOG_matrix)

#read in metadata
swog_meta = read.table("C:/Users/matts/Desktop/projects/spp1/S1314 metadata.txt", sep = "\t", header = TRUE)

#Return DEGs for Uro
test_df = degscore(expression_data = swog_expressions,
                   this_metadata = swog_meta,
                   subtypes = "Uro",
                   return_this = "degs")

}
