% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_degscore.R
\name{get_degscore}
\alias{get_degscore}
\title{Get DEG Score.}
\usage{
get_degscore(
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  this_subtype = NULL,
  exp_data = NULL,
  exclude = NULL,
  out_path = NULL,
  write_degs = FALSE,
  cat_variable = NULL,
  subtype_variable = NULL,
  sample_id_col = NULL,
  return_this = "degs"
)
}
\arguments{
\item{these_sample_ids}{Optional, a vector of multiple sample IDs 
(or a single sample ID as a string) that you want results for.}

\item{these_samples_metadata}{Optional, a metadata table (with sample IDs in 
a column) to subset the return to. If not provided (and if `these_sample_ids` 
is not provided), the function will return all samples.}

\item{this_subtype}{Optional parameter.Allows the user to subset samples to a 
specific subtype of interest. To be used together with `this_variable`.}

\item{exp_data}{Required. A data frame with expression values of interest.}

\item{exclude}{Optional, Lets the user exclude samples of interest. Should be 
a vector with sample IDs available in the metadata.}

\item{out_path}{Required if `write_degs = TRUE`, target path to exported DEGs, 
otherwise disregarded.}

\item{write_degs}{Boolean parameter, default is set to FALSE. Set to TRUE for 
exporting DEGs, if set to TRUE, `out_path` also needs to be specified.}

\item{cat_variable}{Required parameter. This should be the categorical variable 
that is intended for testing. In addition, this should also be a variable of 
type factor, with exactly 2 levels.}

\item{subtype_variable}{Required if `this_subtypes` is defined. Should be the name of 
the column in the metadata holding subtype classification. Otherwise not needed.}

\item{sample_id_col}{Required parameter, should be the name of the column in 
the metadata data frame with sample IDs.}

\item{return_this}{Decides what output is returned, could be one of the 
following; sample_ids, metadata, expressions, eset, design_matrix, degs, all, and nothing}
}
\value{
Data frame with DEGs as well as ranked score.
}
\description{
Return a DEG object with score for each gene.
}
\details{
See parameter descriptions for more info, still wip.
}
\examples{
#load packages
library(dplyr, limma, Biobase)

#Return DEGs for Uro
my_uros = get_degscore(exp_data = swog_exp,
                       these_samples_metadata = swog_meta,
                       this_subtype = "Uro", 
                       subtype_variable = "LundTax2023_simple", 
                       sample_id_col = "ALTPATID",
                       cat_variable = "pT0",
                       return_this = "degs")

}
