% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glmscores.R
\name{get_glmscores}
\alias{get_glmscores}
\title{Get General Linear Model Scores.}
\usage{
get_glmscores(
  these_predictions = NULL,
  these_samples_metadata = NULL,
  subtype_class = "5class",
  this_subtype = NULL,
  cat_variable = NULL,
  row_to_col = FALSE,
  sample_id_col = NULL,
  score_columns = NULL
)
}
\arguments{
\item{these_predictions}{Required parameter. A data frame with predictions
scores, subtype and metadata information.}

\item{these_samples_metadata}{Optional, a data frame with metadata of interest,
function will subset return to what samples are included in this object.}

\item{subtype_class}{Can be one of the following; 5class or 7class. Default is 5class.}

\item{this_subtype}{Optional parameter. Allows the user to subset the return 
to a specific subtype within the selected class. If not specified, the function 
will return a data frame not dependent on subtype.}

\item{cat_variable}{Required parameter. Specify the two level categorical variable you want to test for.}

\item{row_to_col}{Boolean parameter. Set to TRUE to transform row names of the 
metadata to a new column called sample_id. Default is FALSE.}

\item{sample_id_col}{Parameter dictating the column name with sample IDs, 
the function expects this column to be sample_id but the user can override 
this if they know the name for this column.}

\item{score_columns}{Optional parameter. Should be a list with column names 
with signatures of interest + the column sample_id. If not specified, 
the function assumes the signature nomenclature from the generated by 
[LundTax2023Classifier::predict_LundTax2023()] function.}
}
\value{
A data frame with statistical scores for the selected samples/subtypes.
}
\description{
Run a Mann-Whitney and general linear regression on a set of
prediction scores.
}
\details{
This function takes a data frame with prediction data`these_predictions`
and executes statistical tests to find significant prediction scores based on
a categorical variable (i.e response, tumour grade, etc.). Currently, the
function expects the incoming data to be the score output from
[LundTax2023Classifier::predict_LundTax2023()], together with metadata
information of interest (e.g two level categorical) and subtype classification
information. In the future, this function will accept a seperate metadata
object and automagically extract all the other needed information form the
returned list from the previous mentioned function. The user have the option
to point the function to the categorical variable with `cat_variable`.
The return can be further subset by subtype by using the `this_subtype`
variable (one of Uro, BaSq, GU, Mes or ScNE).
}
\examples{
#load packages
library(dplyr, stats)

#run general linear models
my_glm = get_glmscores(these_predictions = swog_pred,
                       these_samples_metadata = swog_meta,
                       subtype_class = "5class",
                       this_subtype = "Uro",
                       cat_variable = "SEX",
                       sample_id_col = "ALTPATID")

}
