% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_predwrangler.R
\name{int_predwrangler}
\alias{int_predwrangler}
\title{Prediction Wrangler}
\usage{
int_predwrangler(
  these_predictions = NULL,
  these_samples_metadata = NULL,
  score_columns = NULL,
  sample_id_col = NULL,
  row_to_col = FALSE,
  cat_variable = NULL,
  subtype_class = "5class",
  this_subtype = NULL
)
}
\arguments{
\item{these_predictions}{Required. Output from [LundTax2023Classifier::predict_LundTax2023()].}

\item{these_samples_metadata}{Required. Metadata associated with he prediction output. 
Also possible for the user to provide a metadata subset with samples of interest, 
the return will be restricted to the samples within the specified group}

\item{score_columns}{Optional parameter. Should be a list with column names 
with signatures of interest + the column sample_id. If not specified, 
the function assumes the signature nomenclature from the generated by 
[LundTax2023Classifier::predict_LundTax2023()] function.}

\item{sample_id_col}{Parameter dictating the column name with sample IDs, 
the function expects this column to be sample_id but the user can override 
this if they know the name for this column.}

\item{row_to_col}{Boolean parameter. Set to TRUE to transform row names of the 
metadata to a new column called sample_id. Default is FALSE.}

\item{cat_variable}{Required parameter. This should be the categorical variable 
that is intended for testing. In addition, this should also be a variable of 
type factor, with exactly 2 levels.}

\item{subtype_class}{Can be one of the following; 5class or 7class. Default is 5class.}

\item{this_subtype}{Optional parameter. Allows the user to subset the return 
to a specific subtype within the selected class. If not specified, the function 
will return a data frame not dependent on subtype.}
}
\value{
A data frame ready for `get_glmscores`.
}
\description{
Internal function called by `get_glmscores`. 
Not meant for out-of-package usage.
}
\details{
Takes an output from [LundTax2023Classifier::predict_LundTax2023()]
and pre-processing it so that statistical tests can be executed on this data.
For more info, see the documentation of `get_glmscores`.
}
\examples{
#load packages
library(dplyr)
  
#run helper                                  
my_out = int_predwrangler(these_predictions = swog_pred, 
                          these_samples_metadata = swog_meta, 
                          subtype_class = "5class", 
                          this_subtype = "Uro", 
                          cat_variable = "SEX", 
                          sample_id_col = "ALTPATID")
                          
}
