% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lund_colors}
\alias{lund_colors}
\title{Lund Colors.}
\format{
A list of 4.
}
\usage{
data(lund_colors)
}
\description{
Colors used for Lund Taxonomy subtypes.
}
\details{
A list with 4 elements.

\itemize{
 \item lund_colors. Vector of characters, 10.
 \item lund_color_transp. Vector of characters, 10.
 \item stage_colors. Vector of characters, 5.
 \item dataset_colors. Vector of characters, 4.
}
}
\keyword{datasets}
