% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glmforest.R
\name{plot_glmforest}
\alias{plot_glmforest}
\title{GLM Forest Plot.}
\usage{
plot_glmforest(
  these_predictions = NULL,
  these_samples_metadata = NULL,
  this_glm = NULL,
  plot_title = "My Plot",
  all_subs = FALSE,
  subtype_class = "5class",
  cat_variable = NULL,
  this_subtype = NULL,
  sample_id_col = NULL,
  row_to_col = FALSE
)
}
\arguments{
\item{these_predictions}{Required parameter if `this_glm` is not provided. 
Should be output from [LundTax2023Classifier::predict_LundTax2023()].}

\item{these_samples_metadata}{Required parameter if `this_glm`is not provided. 
Metadata associated with he prediction output. Also possible for the user to 
provide a metadata subset with samples of interest, the return will be 
restricted to the samples within the specified group}

\item{this_glm}{Required paraemter if `these_predictions` and `these_samples_metadata` 
is not provided. The output from `get_glmscores`.}

\item{plot_title}{Title for plot.}

\item{all_subs}{Boolean, default is FALSE. Set to TRUE to return a forest plot 
for all subtypes within the specified class (`subtype_class`).}

\item{subtype_class}{Can be one of the following; 5class or 7class. Default is 5class.}

\item{cat_variable}{Required parameter if `this_glm`is not provided. 
This should be the categorical variable that is intended for testing. 
In addition, this should also be a variable of type factor, with exactly 2 levels.}

\item{this_subtype}{Required parameter if `this_glm` is not provided, see `get_glmscores`.}

\item{sample_id_col}{Required parameter if `this_glm` is not provided, see `get_glmscores`.}

\item{row_to_col}{Required parameter if `this_glm` is not provided, see `get_glmscores`.}
}
\value{
A forest plot as grub object.
}
\description{
Construct a forest plot using GLM data for a set of signature scores.
}
\details{
This function internally calls `get_glmscore`, which internally 
calls `int_predwrangler`. This happens if the user does not call the plotting 
function using the `this_glm` parameter. Many of the parameters in this 
function are recycled in the internal funciton calls to improve user flexibility.
The only parameters that are directly called by this function is; `plot_title`,
`all_subs` and `this_glm`. It's possible to return a forest plot for all 
subtypes within the specified `subtype_class`. To do so, set `all_subs = TRUE`. 
This will trigger a loop for which `get_glmscore` is internally called for each subtype.
To return a plot for just one subtype, set `all_subs = FALSE` and specify the 
subtype and class with `this_subtype` and `subtype_class`. To return a plot 
for more than one subtype, but not all, call get_glmscore for each itteration 
of subtypes, see last example in the docs.
}
\examples{
#load pacakges
library(dplyr, ggplot2)

#plot using prediction and metadata, Uro
plot_glmforest(these_predictions = swog_pred,
               these_samples_metadata = swog_meta,
               plot_title = "Uro, pT0",
               all_subs = FALSE,
               subtype_class = "5class", 
               this_subtype = "Uro",
               cat_variable = "pT0",
               sample_id_col = "ALTPATID")#'

#plot all subtypes within the 7 class
plot_glmforest(these_predictions = swog_pred,
               these_samples_metadata = swog_meta,
               plot_title = "All 7class",
               all_subs = TRUE,
               subtype_class = "7class", 
               cat_variable = "pT0",
               sample_id_col = "ALTPATID")#'

#plot Uro and GU
#get Uro glms
uro_glm = get_glmscores(these_predictions = swog_pred,
                        these_samples_metadata = swog_meta,
                        subtype_class = "5class",
                        this_subtype = "Uro",
                        cat_variable = "pT0",
                        sample_id_col = "ALTPATID")#'

#get GU glms
gu_glm = get_glmscores(these_predictions = swog_pred,
                       these_samples_metadata = swog_meta,
                       subtype_class = "5class",
                       this_subtype = "GU",
                       cat_variable = "pT0",
                       sample_id_col = "ALTPATID")#'

#draw plot
plot_glmforest(this_glm = rbind(uro_glm, gu_glm))

}
