% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores.R
\name{plot_scores}
\alias{plot_scores}
\title{Plot Signature Scores.}
\usage{
plot_scores(
  this_data = NULL,
  out_path = NULL,
  out_format = "png",
  return_scores = FALSE,
  to_xlsx = FALSE,
  title = NULL,
  hm_split = NULL,
  hm_cluster = FALSE,
  plot_anno_legend = NULL,
  plot_hm_legend = FALSE,
  plot_width = 14,
  plot_height = 6,
  plot_font_size = 12,
  verbose = TRUE
)
}
\arguments{
\item{this_data}{A list with a data frame object called scores. Returned with `predict_LundTax2023`.}

\item{out_path}{Optional, set path to export plot. If not provided, tidy version of incoming
scores in data frame format will be returned (`return_scores` will be auto-defaulted to TRUE).}

\item{out_format}{Required parameter if `out_path` is specified. Can be "png" (default) or "pdf".
if pdf, the returned pdf will be in A4 (horizontal) format, if png is specified the user can
control the dimensions, see `plot_width` and `plot_height`.}

\item{return_scores}{Set to TRUE to return prediction scores in a tidy format. Default is FALSE.}

\item{to_xlsx}{Boolean parameter, set to TRUE to export score data frame in xlsx format. Default is FALSE.
If set to TRUE, the spreadsheet will be saved to the same path as the heatmap.}

\item{title}{Required parameter. Heatmap title, will also be pasted to the exported file(s) as well
as a new column in the scores data frame under cohort.}

\item{hm_split}{Optional parameter for controlling how the data is split into different groups.
If not provided, the function will split on subtype.}

\item{hm_cluster}{Boolean parameter, set to TRUE to cluster the rows (default is FALSE).}

\item{plot_anno_legend}{Expects a vector with TRUE/FALSE (7 in total), thsi decides what legends
will be on the final heatmap. Default is to only show the legend for the subtypes.}

\item{plot_hm_legend}{Boolean parameter. Set to TRUE to show heatmap legend. Default is FALSE.}

\item{plot_width}{If `out_format = "png"`, this parameter controls the width in pixels.
Default is 14 inches (4200 pixels at 300 PPI).}

\item{plot_height}{If `out_format = "png"`, this parameter controls the height in pixels.
Default is 6 inches (1800 pixels at 300 PPI)}

\item{plot_font_size}{Optional parameter to control the size of the font in the generated heatmap.
Note, the title of the plot will always be twice that of the set value here (default = 12).}

\item{verbose}{Set to TRUE for debugging purposes. Default is FALSE.}
}
\value{
Data frame with prediction score for each sample and class, if return_scores = TRUE.
Otherwise, nothing.
}
\description{
Build a heatmap with scores retrieved with the predict_LundTax2023 function.
}
\details{
Construct and export (pdf or png) a highly customizable heatmap visualizing prediction
scores for each sample and class, predicted with `predict_LundTax2023`.
This function depends on Complexheatmap. It is also possible to return a data frame with
prediction scores in a tidy format. To do so, set `return_scores = TRUE`. For a greater explanation
on how to use the function, see parameter descriptions and examples.
}
\examples{
\dontrun{
#load pacakges
library(ComplexHeatmap, ggplot2, grid, circlize, openxlsx, grDevices, utils)

plot_scores(this_data = swog_pred,
            out_path = "../",
            out_format = "pdf",
            title = "Lund2017")
}

}
