% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swog_exp}
\alias{swog_exp}
\title{SWOG Expressions.}
\format{
A data frame with 15986 rows (genes) and 163 columns (samples).
}
\usage{
data(swog_exp)
}
\description{
Example gene expression dataset. Obtained from the Southwest Oncology Group (SWOG).
doi: 10.1158/1078-0432.CCR-23-0602.
}
\details{
A data frame with samples in columns and expression values in rows with gene
names (HGNC symbols) as row names.

\itemize{
}
}
\keyword{datasets}
