% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swog_meta}
\alias{swog_meta}
\title{SWOG Metadata.}
\format{
A data frame with 163 rows (samples) and 39 columns (metadata).
}
\usage{
data(swog_meta)
}
\description{
Metadata associated with the bundled expression dataset.Obtained from the
Southwest Oncology Group (SWOG). doi: 10.1158/1078-0432.CCR-23-0602.
}
\details{
A data frame with samples in rows and metadata variables in the columns.

\itemize{
 \item ALTPATID. Unique sample ID.
 \item ARMNAME.
 \item STRAT1.
 \item STRAT2.
 \item AGE.
 \item SEX.
 \item RACE.
 \item prim_analysis.
 \item itt_analysis.
 \item adequate_tiss.
 \item ANY_TRT.
 \item GFS.
 \item adewuate_chemo.
 \item CYSTECTOMY.
 \item PATHSTHT.
 \item PATHSTGN.
 \item PATHSTGM.
 \item NLYMPHN.
 \item NPOSNOD.
 \item PATH_RESP.
 \item pT0.
 \item OS_IND.
 \item OS_TIM.
 \item PFS_IND.
 \item PFS_TIM.
 \item TT_offtx.
 \item TT_cyst.
 \item GC_FAV.
 \item ddMVAC_FAV.
 \item GC_FAV_dup.
 \item ddMVAC_FAV_dup.
 \item GEO.
 \item tchagrp.
 \item tcga.
 \item LundTax2023_simple.
 \item consensus.
 \item mda.
 \item mdagrp.
 \item poor_quality.
}
}
\keyword{datasets}
