% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swog_pred}
\alias{swog_pred}
\title{SWOG Predictions.}
\format{
A list of 5.
}
\usage{
data(swog_pred)
}
\description{
Prediction output from running LundTax2023_predict on the SWOG expression set.
}
\details{
A list of 5.

\itemize{
 \item data. Expression data.
 \item subtype_scores. Prediction scores related to subtypes.
 \item prediction_7classes. Consensus subtype predictions for 7 class.
 \item prediction_5classes. Consensus subtype predictions for 5 class.
 \item scores. Signature scores.
}
}
\keyword{datasets}
