% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{UroBladR-package}
\alias{UroBladR}
\alias{UroBladR-package}
\title{isoploteR
Takes an expression matrix and subsets to sample IDs of interest for the selected genes/isoforms.
Additionally, the function can also return a basic box plot with the returned data.}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Adam Mattsson \email{adam.mattsson@med.lu.se} (\href{https://orcid.org/0000-0002-6318-7912}{ORCID})

}
