% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isos.R
\name{get_isos}
\alias{get_isos}
\title{Get Isos}
\usage{
get_isos(
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  this_data,
  annotations,
  my_genes,
  these_isoforms,
  verbose = TRUE,
  to_frac = TRUE,
  return_all = FALSE,
  return_plot = TRUE,
  plot_title = "My Plot",
  plot_subtitle = "My subtitle",
  my_color = "#499851"
)
}
\arguments{
\item{these_sample_ids}{Optional, a vector of multiple sample IDs (or a single sample ID as a string)
that you want results for.}

\item{these_samples_metadata}{Optional, a metadata table (with sample IDs in a column) to subset
the return to. If not provided (and if `these_sample_ids` is not provided), the function will
return all samples.}

\item{this_data}{Optional parameter for specifying the expression matrix of interest.
If not provided, the function will use a bundled data object for this purpose.}

\item{annotations}{Gene annotations. A data frame with txID, GeneID, and Symbol as columns.}

\item{my_genes}{Optional. A list with genes of interest. If provided the function will subset to
all isoforms for the selected gene(s)}

\item{these_isoforms}{Required if not `my_genes` is provided. A vector of characters with isoforms
of interest in Gene ID format.}

\item{verbose}{Boolean parameter. Set to FALSE to minimize output to console. Default is TRUE.}

\item{to_frac}{Boolean parameter. If TRUE (default), each fraction per isoform for each sample
will be reported. If `return_plot` is set to TRUE, this parameter will auto-default to TRUE.
To get back TPM values in the data frame, first set `return_plot = FALSE`.}

\item{return_all}{Boolean parameter. Set to TRUE to return use all sample IDs in the provided
dataset with `this_data`. Default is FALSE.}

\item{return_plot}{Boolean parameter. Set to TRUE (default) to return box plot for selected
isoforms and samples.}

\item{plot_title}{Optional parameter for naming the returned plot (if return_plot = TRUE).}

\item{plot_subtitle}{Optional parameter for adding a subtitle to the returned plot
(if return_plot = TRUE).}

\item{my_color}{}
}
\value{
A data frame with gene IDs and the corresponding GEX as rows and sample IDs in the columns.
}
\description{
Subset a data frame with expression values for different genes and isoforms to only
include a set number of genes/isoforms for selected sample IDs.
}
\details{
Takes an expression matrix and subsets to sample IDs of interest for the selected genes/isoforms.
Additionally, the function can also return a basic box plot with the returned data.
To ensure compatibility (i.e column names and data types) of annotations and expression matrix format,
please see the bunlded data object.
}
\examples{
#get samples
my_samples = colnames(expression_sub)[-1]

#run function
these_isos = get_isos(this_data = expression_sub,
                      annotations = gene_annotations,
                      these_sample_ids = my_samples,
                      these_isoforms = c("ENST00000395080",
                                         "ENST00000237623",
                                         "ENST00000360804",
                                         "ENST00000508233",
                                         "ENST00000681973"),
                      plot_title = "SPP1",
                      plot_subtitle = "Isoforms Frequency")

}
